Update lab to scaffold tasks
